split off s3 connection logic
